fix and parallelize eiwres
Implements a hook for _derived to call simobs on a population and just get the derived appropriately calculated and returned, and then uses this population-based _derived in eiwres. However, I also noticed the eiwres had a bug where it wasn't actually using the new distributions each time, so this also fixes the bug.
parallel _derived