simobs multithreaded reproducibility by default
Fixes https://github.com/PumasAI/Pumas.jl/issues/817 by making the seed default to being based on the hash of the ID. This makes everything reproducible when parallel. This default can be disabled by `obsseed = nothing`, in which case the global rng takes hold and randomness ensues. This does have the unfortunate drawback that randomness is tied to the ID, so if all subjects have the same ID they will have the same randomness. I don't think this is too weird though, since ID is one of the required arguments in a Subject so it shouldn't show up... but it does show up in the diagnostics since we just create nsim copies of the same subject, and so in that case it's set to `obsseed=nothing` for full randomness.